Rename NuGetUpgrader namespace and move NuGetUpgrader unit tests to NuGetUpgrade folder
