+++ GVFS/GVFS.UnitTests/Common/InstallManifestTests.cs
@@ -1,5 +1,4 @@
-﻿using GVFS.Common;
-using GVFS.Common.NuGetUpgrader;
+﻿using GVFS.Common.NuGetUpgrade;
 using GVFS.Tests.Should;
 using Newtonsoft.Json;
 using NUnit.Framework;
